Create hook for getting default nonce
